This fixes an error where eigen couldn't transpose an nx1 matrix.
This was due to the forced row major layout which is not default in
Eigen. I do not believe there is any reason to use row major layout.
Possibly this was done for compatibility with the arm CMSIS library
but that library is never used in tandem with the eigen library.